clippy: address workspace-wide lints to unblock CI
The rustfmt+clippy CI job was failing on ~90 clippy errors across 15
crates (manual_memcpy, needless_range_loop, type_complexity, etc.).
Most were auto-fixable via `cargo clippy --fix`; the rest were resolved
manually — slice::copy_from_slice for manual copies, iter::enumerate for
index-only loops, `.clamp(a, b)` for manual min/max chains, struct
literal init for field_reassign_with_default, and crate-local allows for
signal-processing signatures where the argument count is spec-driven.

Test-file lints in signal-processing roundtrips and an always-true
assertion in g729 (s.abs() <= i16::MAX) were allowed or removed where
the readability cost of "fixing" them exceeded the lint's value.

`cargo clippy --workspace --all-targets -- -D warnings` and
`cargo test --workspace --lib` both pass clean locally (974/974).

+++ crates/oxideav-iff/src/svx.rs
@@ -466,7 +466,7 @@ impl Muxer for SvxMuxer {
         }
         // IFF chunks pad to even length; BODY is the last child chunk so
         // its pad byte (if any) also pads the enclosing FORM.
-        let body_pad = (self.body_bytes & 1) as u64;
+        let body_pad = self.body_bytes & 1;
         if body_pad == 1 {
             self.output.write_all(&[0u8])?;
         }
